requirements.txt
requests>=2.31.0
pyyaml>=6.0
streamlit>=1.32.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
line-bot-sdk>=3.0.0
pandas>=2.0.0
